Moving Audit Log listing to its own module with its own role protections outside of Admin
