+++ src/app/core/audit/list-audit-entries/list-audit-entries.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../common/paging/quick-filters/quick-filters.component';
+@import '../../../common/paging/quick-filters/quick-filters.component';
 
 .audit-filter {
