+++ src/app/core/core.module.ts
@@ -9,6 +9,7 @@ import { PopoverModule } from 'ngx-bootstrap/popover';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 
 import { AdminModule } from './admin/admin.module';
+import { AuditModule } from './audit/audit.module';
 import { HelpModule } from './help/help.module';
 import { FeedbackModule } from './feedback/feedback.module';
 
@@ -59,6 +60,7 @@ export function getConfiguration(configService: ConfigService) {
 		TooltipModule,
 
 		AdminModule,
+		AuditModule,
 		HelpModule,
 		FeedbackModule,
 		TeamsModule,
